Support for specifying TVM IR in Python.
The big new user-facing module is torch.isl, which exposes
IslFunction, a function which can be subclassed with a TVM IR
description, and then on-the-fly compiles it into CUDA code.
See IslMatMult for an example of how to use it.

There is also a bugfix where THPUtils_unpackLong in python_numbers.h
is given a different name so it doesn't conflict with the similarly
named macro (causing very strange errors).  It will be worth working
out exactly what is going on here.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>