Standalone linear test.
This is a particularly interesting E2E test, because when I ran
linux-64/caffe2-0.8.dev-py27hf340d0f_1.tar.bz2 (-c ezyang)
on superbox (PyTorch build machine), it didn't segfault but
gave incorrect results when running an FC layer on GPU.
Recompiling protobuf and caffe2 on the machine itself solved
the problem, so it is likely to be some subtle ABI incompatibility
that wasn't caught during linking.

I'd kind of like to investigate some more... but no time.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>