Update to ToffeeIR HEAD, minus BC-breaking MS changes.
This commit advances the ToffeeIR submodule pointer to
pr/hot-fixed-api, which is a TEMPORARY branch that tracks
ToffeeIR HEAD, but with the Microsoft breaking patches reverted.
This means we can develop against the new ToffeeIR API until
we can handle the MS changes, and then we can swing the pointer
back to master.

The code changes indicate fixes for namedtuple API change.
We'll need to preserve those.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>